Managed to save all details records in one file

+++ breaches.py
@@ -1,6 +1,6 @@
 import os
-import time
 import urllib.error
+from os.path import exists
 
 import pandas
 import requests
@@ -19,7 +19,7 @@ def scrape_breaches():
     data = pandas.DataFrame(res,
                             columns=['case/breach', 'defendant\'s_name', 'hearing_date', 'result',
                                      'fine £', 'act_or_regulation'])
-    data['cytora_ingest_ts'] = time.time()
+    data['cytora_ingest_ts'] = pandas.to_datetime('today')
 
     try:
         os.makedirs('Breaches')
@@ -43,7 +43,7 @@ def scrape_breaches():
             data = pandas.DataFrame(res,
                                     columns=['case/breach', 'defendant\'s_name', 'hearing_date', 'result',
                                              'fine £', 'act_or_regulation'])
-            data['cytora_ingest_ts'] = time.time()
+            data['cytora_ingest_ts'] = pandas.to_datetime('today')
             data.to_csv('Breaches/breaches.csv', index=False, mode='a', header=False)
 
             i += 1
@@ -122,10 +122,14 @@ def fetch_breaches_details(case_id: str, breach_number: str, endpoint: str) -> d
             ]
         )
 
-        new_df['cytora_ingest_ts'] = time.time()
+        new_df['cytora_ingest_ts'] = pandas.to_datetime('today')
 
-        file_path = f'Breaches/{endpoint}.csv'
-        new_df.to_csv(file_path, index=False)
+        file_path = f'Breaches/breaches_details.csv'
+
+        if not exists(file_path):
+            new_df.to_csv(file_path, index=False)
+        else:
+            new_df.to_csv(file_path, index=False, mode='a', header=False)
 
         return {
             'state': True,
